Implement bookmarkers script
The script adds a new icon underneath the bookmarks UI for a question, when
there is at least 1 bookmark on that post. When clicked, the most recent known
list of users that bookmarked the question is fetched from the Stack Exchange
data explorer, after which user details are fetched from the Stack Exchange API
and displayed in a popup box.