fix: unbreak automatic updates for userscripts
I misunderstood what downloadURL is for when also setting the updateURL
header: it's not meant to point to the place where this version was
downloaded. Instead, it is meant to serve the new version of the script
when the headers-only updateURL states there is a new version to be
found.

Drop the updateURL header for now, and point downloadURL to main/dist.

See #533.