fix(connected-users): simplify div size discovery
The ensureHasSize function doesn't need to meddle with s-expandable, as
became evident when I was testing the function; it never expanded any
expandables. The _width_ of elements is not affected by collapsing them,
so we can eliminate handling these safely.

The utilities have been moved to the utils module and are now tested.